requirements.txt
asgiref==3.2.7
bootstrap4==0.1.0
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cssselect2==0.3.0
cycler==0.10.0
Cython==0.29.16
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.7
django-ckeditor==5.9.0
django-cleanup==4.0.0
django-crispy-forms==1.9.1
django-js-asset==1.2.2
django-template-editor==0.1.1
django-template-field==0.3.1
html5lib==1.0.1
idna==2.9
joblib==0.14.1
networkx==2.4
Pillow==7.1.2
pip==20.1.1
pycparser==2.20
pyparsing==2.4.7
Pyphen==0.9.5
python-dateutil==2.8.1
pytz==2019.3
setuptools==46.0.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
tinycss2==1.0.2
urllib3==1.25.9
webencodings==0.5.1
wheel==0.34.2